backend/requirements.txt
flask
flask-cors
matplotlib
numpy
pandas
scikit-learn
python-dotenv
pymongo